migration(reports): migrate reports to tsq and split refactor reports.tsx
- fix(#609): json parsing
- fix missing key in Diary
- hide recharts erorrs
- remove api call from services
- remove service file
