+++ SparkyFitnessFrontend/src/constants/reports.ts
@@ -1,0 +1,9 @@
+// Body Battery metric names as they come from Garmin sync
+export const BODY_BATTERY_METRICS = [
+  'Body Battery Highest',
+  'Body Battery Lowest',
+  'Body Battery At Wake',
+  'Body Battery Charged',
+  'Body Battery Drained',
+  'Body Battery Current',
+];
