+++ SparkyFitnessFrontend/src/pages/Reports/FastingReport.tsx
@@ -28,9 +28,7 @@ interface FastingReportProps {
 
 const COLORS = ['#6366f1', '#06b6d4', '#f59e0b', '#ef4444'];
 
-export const FastingReport: React.FC<FastingReportProps> = ({
-  fastingData,
-}) => {
+export const FastingReport = ({ fastingData }: FastingReportProps) => {
   const { t } = useTranslation();
   const { formatDateInUserTimezone } = usePreferences();
   const [isMounted, setIsMounted] = useState(false);
