+++ SparkyFitnessFrontend/src/pages/Reports/NutritionChartsGrid.tsx
@@ -23,27 +23,7 @@ import {
 import type { UserCustomNutrient } from '@/types/customNutrient';
 import { CENTRAL_NUTRIENT_CONFIG } from '@/constants/nutrients';
 import { formatNutrientValue } from '@/utils/nutrientUtils';
-interface NutritionData {
-  date: string;
-  calories: number;
-  protein: number;
-  carbs: number;
-  fat: number;
-  saturated_fat: number;
-  polyunsaturated_fat: number;
-  monounsaturated_fat: number;
-  trans_fat: number;
-  cholesterol: number;
-  sodium: number;
-  potassium: number;
-  dietary_fiber: number;
-  sugars: number;
-  vitamin_a: number;
-  vitamin_c: number;
-  calcium: number;
-  iron: number;
-  [key: string]: number | string; // Add index signature for custom nutrients
-}
+import { NutritionData } from '@/services/reportsService';
 
 interface NutritionChartsGridProps {
   nutritionData: NutritionData[];
