+++ SparkyFitnessFrontend/src/pages/Reports/SpO2Gauge.tsx
@@ -1,4 +1,4 @@
-import type React from 'react';
+import { getSpO2StatusInfo } from '@/utils/reportUtil';
 import { useTranslation } from 'react-i18next';
 
 interface SpO2GaugeProps {
@@ -7,56 +7,7 @@ interface SpO2GaugeProps {
   strokeWidth?: number; // px, default 12
 }
 
-const getSpO2StatusInfo = (
-  value: number
-): { status: string; color: string; description: string } => {
-  if (value < 70) {
-    return {
-      status: 'Critical',
-      color: '#ef4444',
-      description: 'Dangerously low oxygen levels. Seek medical attention.',
-    };
-  } else if (value < 80) {
-    return {
-      status: 'Low',
-      color: '#f97316',
-      description: 'Below normal oxygen levels. Monitor closely.',
-    };
-  } else if (value < 90) {
-    return {
-      status: 'Moderate',
-      color: '#eab308',
-      description: 'Slightly below optimal levels.',
-    };
-  } else if (value < 95) {
-    return {
-      status: 'Normal',
-      color: '#22c55e',
-      description: 'Healthy oxygen saturation levels.',
-    };
-  } else {
-    return {
-      status: 'Excellent',
-      color: '#22c55e',
-      description: 'Optimal oxygen saturation.',
-    };
-  }
-};
-
-// Get color for a specific SpO2 value (for bar chart)
-// eslint-disable-next-line react-refresh/only-export-components
-export const getSpO2Color = (value: number): string => {
-  if (value < 70) return '#ef4444';
-  if (value < 80) return '#f97316';
-  if (value < 90) return '#eab308';
-  return '#22c55e';
-};
-
-const SpO2Gauge: React.FC<SpO2GaugeProps> = ({
-  value,
-  size = 160,
-  strokeWidth = 12,
-}) => {
+const SpO2Gauge = ({ value, size = 160, strokeWidth = 12 }: SpO2GaugeProps) => {
   const { t } = useTranslation();
   const { status, color, description } = getSpO2StatusInfo(value);
